CIRC-519 Rename classes
